Remove one more unneccesary line of config - @PropertySource("classpath:application.properties")
Also changed properties file and added @GeneratedValue to both the
Entities to tell JPA(or hibernate) to create table only if it doesn't
exist.

Also in last commit, I forgot to mention that I changed the book table
name to BookJPA2 from BookJPA, as it was giving some error.

And remember, to create first Book, there must be an author present and
u have to create book using one of the Ids of author as AuthorId, Coz of
Foreign key constraint.